Modules are referenced and unreferenced on tab change. This behaviour should improve the performance since before all modules were loaded at initialization.
The Red5 connection is closed on each tab change because the Flash GCollector is unpredictable.
Still I've found a bug on the Evaluation module that causes multiple NetStreams to play simultaneously
as if some remain of the previous instance was there.

Module Deeplinking works, but action and target fragments are still to be done/fixed to work.